Fix linter issues in the OpsRamp code
Brings the OpsRamp packages in line with the repository lint rules.

- Extend the revive var-naming allowlist with UUID, UID, JSON and SSH.
  These names mirror the JSON of the API being wrapped, and the repository
  already takes this approach for ID, URL, API, IDS and IP. It avoids
  renaming sixteen exported struct fields for a cosmetic gain.
- Drop redundant []byte conversions where the value is already []byte.
- Rename underscore-style locals, and move one declaration inside the
  block that uses it, which also removes an ineffectual assignment.
- Replace the deprecated net.Error.Temporary call. A net.Error which is
  not a timeout now falls through to the same text check as any other
  error, which it previously bypassed.
- Replace the deprecated strings.Title with a Caser, constructed per call
  since a Caser is not safe for concurrent use.
- Type the client id as int64 so it maps onto the Terraform attribute
  without an unprovable conversion.
- Annotate the four request bodies that necessarily carry credentials.
- Skip rather than fail when acceptance-test environment variables are
  absent, so go test succeeds for anyone not targeting a live tenant.
  All OpsRamp packages now pass go test without credentials.

Formatting fixes from gofumpt, goimports and nlreturn are included.
Line-length is addressed separately.

+++ opsramp/client/client_models.go
@@ -23,7 +23,9 @@ type CreateClient struct {
 
 // ClientResponse represents the API response for a client
 type ClientResponse struct {
-	Id                uint     `json:"id"`
+	// Id is int64 rather than uint so it maps directly onto the Terraform
+	// Int64 attribute without a conversion that cannot be proven safe.
+	Id                int64    `json:"id"`
 	UniqueId          string   `json:"uniqueId"`
 	Name              string   `json:"name"`
 	Activated         bool     `json:"activated"`
